Client logic finalization
Added function for text formatting.
Split roll's logic into two more functions, one for animating the slots and one for handling the roll data.
Adding credentials to start game to reuse the session's values.
Adding color for message based on if the roll is winning.
Added credentials to cashout's fetch.
Added function to toggle buttons.
Added function to set slots to a value